Removed all 8 jbpm Gateway FeatureContainer classes because the only thing they did was set the name to blank - this can be handled generically in Jbpm5RuntimeExtension.modelObjectCreated().
Moved BoundaryEventPositionHelper from UI to core plugin - this will be
needed to fix the connection routing problem for BoundaryEvents

Fixed an NPE in ConnectionFeatureContainer when trying to update a
Message Flow that connects a Participant Band to a Message (a so-called
"ChoreographyMessageLink")

Changed AddBendpointFeature to prevent adding bendpoints to
ChoreographyMessageLinks

Updated AnchorUtil.findNearestBoundaryAnchor() for BoundaryEvents - only
returns anchors that are outside of the parent shape

Added the jbpm extension attributes to <modelEnablement> list because
all extension elements are now disabled by default.

Reorganized jbpm Task FeatureContainer classes because images were not
being drawn on the specialized Task figures (ManualTask, ServiceTask,
etc.)

Fixed a Folder/IProject class cast exception in NewJbpmProcessWizard

Added missing "isForCompensation" attribute to all Activity property
sheets

Fixed the "Create Operation" functionality for ServiceTask - this should
not have been a RootElementFeature.

Added a "Create Message" button to SendTask and ReceiveTask property
sheets

Fixed the Schema Selection dialog and Schema Object Editor - these did
not handle Java imports correctly

Changed the CreateFeature descriptions for ALL features - the
description text appears in the "Undo" menu and should only be something
like "Create SomeObject", not a verbose description of what a
"SomeObject" is.