Bug 471357: target runtime refactoring
Goal of the refactoring was getting rid of the static
TargetRuntime.currentRuntime, which caused problems.

Extended model resource adapters with an additional
TargetRuntimeAdapter, which provides access to the current runtime.
The runtime is loaded into the resource adapter during initialization.
This way it's possible to access the runtime also through model objects
(using the adapter of it's resource) and diagram objects (using the
linked model object).
Extended API with a TargetRuntime parameter, if there was no otherway to
access it. In some places access the runtime through the editor or it's
behavior.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=471357
Signed-off-by: Flavio Donzé <flavio.donze@gmx.ch>