https://bugs.eclipse.org/bugs/show_bug.cgi?id=464052 Adding a 3rd Pool element is not working correctly when Full-Profile is active
Fix conversions of Process and Collaboration Diagrams when DND a Pool
into a Process Diagram, or when DND an Activity (or Event or Gateway)
onto canvas in a Collaboration Diagram.

Changed ModelHandler and Bpmn2ModelerFactory to look more like the Mars
version: different usages of create() and createObject() were confusing,
we we now have:
 * create() is the low-level object creation mechanism that invokes
ModelDecorator.populateObject() to allow the TargetRuntime to decorate
the object during creation.
 * createObject() routes model object creation through the
ExtendedPropertiesAdapter.createObject() for the object for additional
object/context initialization.

These bugs were discovered while fixing:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=488792
Tool Palette extension point not working as expected